Move code in the `mobile` directory
